Moved setup to accommodate test running leaner
